LegacyPasswds dump/restore
